Preamble section in General Tab is now located in its own Tab "Preamble"
