+++ src/main/java/org/jabref/gui/libraryproperties/LibraryPropertiesViewModel.java
@@ -6,6 +6,7 @@ import org.jabref.gui.libraryproperties.constants.ConstantsPropertiesView;
 import org.jabref.gui.libraryproperties.contentselectors.ContentSelectorView;
 import org.jabref.gui.libraryproperties.general.GeneralPropertiesView;
 import org.jabref.gui.libraryproperties.keypattern.KeyPatternPropertiesView;
+import org.jabref.gui.libraryproperties.preamble.PreamblePropertiesView;
 import org.jabref.gui.libraryproperties.saving.SavingPropertiesView;
 import org.jabref.model.database.BibDatabaseContext;
 
@@ -19,7 +20,8 @@ public class LibraryPropertiesViewModel {
                 new SavingPropertiesView(databaseContext),
                 new KeyPatternPropertiesView(databaseContext),
                 new ConstantsPropertiesView(databaseContext),
-                new ContentSelectorView(databaseContext)
+                new ContentSelectorView(databaseContext),
+                new PreamblePropertiesView(databaseContext)
         );
     }
