+++ src/main/java/org/jabref/gui/libraryproperties/general/GeneralProperties.fxml
@@ -3,7 +3,6 @@
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ComboBox?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextArea?>
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.control.Tooltip?>
 <?import javafx.scene.layout.ColumnConstraints?>
@@ -88,6 +87,4 @@
         <ComboBox fx:id="databaseMode" prefWidth="150.0"
                   GridPane.columnIndex="1" GridPane.hgrow="ALWAYS" GridPane.rowIndex="6"/>
     </GridPane>
-    <Label styleClass="sectionHeader" text="%Preamble"/>
-    <TextArea fx:id="preamble" minHeight="100" prefHeight="180" VBox.vgrow="ALWAYS"/>
 </fx:root>
