+++ src/main/java/org/jabref/gui/libraryproperties/general/GeneralPropertiesViewModel.java
@@ -4,8 +4,6 @@ import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 
-import javax.swing.undo.UndoManager;
-
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ListProperty;
 import javafx.beans.property.ObjectProperty;
@@ -18,7 +16,6 @@ import javafx.collections.FXCollections;
 
 import org.jabref.gui.DialogService;
 import org.jabref.gui.libraryproperties.PropertiesTabViewModel;
-import org.jabref.gui.undo.UndoablePreambleChange;
 import org.jabref.gui.util.DirectoryDialogConfiguration;
 import org.jabref.logic.l10n.Encodings;
 import org.jabref.logic.shared.DatabaseLocation;
@@ -37,7 +34,6 @@ public class GeneralPropertiesViewModel implements PropertiesTabViewModel {
     private final StringProperty generalFileDirectoryProperty = new SimpleStringProperty("");
     private final StringProperty userSpecificFileDirectoryProperty = new SimpleStringProperty("");
     private final StringProperty laTexFileDirectoryProperty = new SimpleStringProperty("");
-    private final StringProperty preambleProperty = new SimpleStringProperty("");
 
     private final DialogService dialogService;
     private final PreferencesService preferencesService;
@@ -45,12 +41,10 @@ public class GeneralPropertiesViewModel implements PropertiesTabViewModel {
     private final BibDatabaseContext databaseContext;
     private final MetaData metaData;
     private final DirectoryDialogConfiguration directoryDialogConfiguration;
-    private final UndoManager undoManager;
 
-    GeneralPropertiesViewModel(BibDatabaseContext databaseContext, DialogService dialogService, PreferencesService preferencesService, UndoManager undoManager) {
+    GeneralPropertiesViewModel(BibDatabaseContext databaseContext, DialogService dialogService, PreferencesService preferencesService) {
         this.dialogService = dialogService;
         this.preferencesService = preferencesService;
-        this.undoManager = undoManager;
         this.databaseContext = databaseContext;
         this.metaData = databaseContext.getMetaData();
 
@@ -68,8 +62,6 @@ public class GeneralPropertiesViewModel implements PropertiesTabViewModel {
         generalFileDirectoryProperty.setValue(metaData.getDefaultFileDirectory().orElse("").trim());
         userSpecificFileDirectoryProperty.setValue(metaData.getUserFileDirectory(preferencesService.getFilePreferences().getUser()).orElse("").trim());
         laTexFileDirectoryProperty.setValue(metaData.getLatexFileDirectory(preferencesService.getFilePreferences().getUser()).map(Path::toString).orElse(""));
-
-        preambleProperty.setValue(databaseContext.getDatabase().getPreamble().orElse(""));
     }
 
     @Override
@@ -101,12 +93,6 @@ public class GeneralPropertiesViewModel implements PropertiesTabViewModel {
         }
 
         databaseContext.setMetaData(newMetaData);
-
-        String newPreamble = preambleProperty.getValue();
-        if (!databaseContext.getDatabase().getPreamble().orElse("").equals(newPreamble)) {
-            undoManager.addEdit(new UndoablePreambleChange(databaseContext.getDatabase(), databaseContext.getDatabase().getPreamble().orElse(null), newPreamble));
-            databaseContext.getDatabase().setPreamble(newPreamble);
-        }
     }
 
     public void browseGeneralDir() {
@@ -154,9 +140,5 @@ public class GeneralPropertiesViewModel implements PropertiesTabViewModel {
 
     public StringProperty laTexFileDirectoryProperty() {
         return this.laTexFileDirectoryProperty;
-    }
-
-    public StringProperty preambleProperty() {
-        return this.preambleProperty;
     }
 }
